pyproject.toml
[project]
name = "langgraphalfred"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "brave-search>=0.1.8",
    "google-api-python-client>=2.176.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "googlemaps>=4.10.0",
    "langchain-core>=0.3.68",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "openwakeword>=0.4.0",
    "playwright>=1.53.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
    "python-weather>=2.1.0",
    "pytz>=2025.2",
    "realtimestt>=0.3.104",
    "websockets>=15.0.1",
]
